+++ src/test/converter2/behavior/externalSymbols.ts
@@ -1,0 +1,5 @@
+/**
+ * Testing custom external link resolution
+ * {@link !Promise}
+ */
+export type P = Promise<string>;